Bug 12836 - Add script to rewrite SRU Explain file
The standard SRU Explain file provided with Koha is static and out of
date. This patch adds a script in misc/maintenance/generate_explain.pl
that can generate a new Explain file, based on Koha's config files and
an optional extra config file. See the POD in the script for more
details.

This is a preview, NOT READY FOR SIGNOFF.