Bug 18585 - Connect to RDF triplestore
This commit adds a 'triplestore' method to C4::Context. It takes
a model name as a parameter and returns a RDF::Trine::Model object
if triplestore.yaml is properly configured.

(At the moment, it just supports RDF::Trine::Store::SPARQL,
but it would be trivial to add support for RDF::Trine::Store::DBI too,
or any other RDF::Trine::Store::* backend.)

This code will provide a base from which people can use RDF::Trine
for querying and updating a RDF triplestore.